+++ R/glm_CMP_methods.R
@@ -1,6 +1,6 @@
 #'AIC and BIC for COM-Poisson Fitted Models
 #'
-#'Computes the Akaike's information criterion or the Shwarz's Bayesian criterion
+#'Computes the Akaike's information criterion or the Schwarz's Bayesian criterion
 #'for COM-Poisson Fits
 #'
 #'@param object an object of class \code{"glm_CMP"}, typically the result of a
@@ -12,10 +12,13 @@ NULL
 #' @rdname AIC_CMP
 #' @export
 #' @examples
+#' ## Estimate a COM-Poisson model
 #' Bids$size.sq <- Bids$size ^ 2
 #' fit <- glm.CMP(formula.mu = numbids ~ leglrest + rearest + finrest +
 #'               whtknght + bidprem + insthold + size + size.sq + regulatn,
 #'               formula.nu = numbids ~ 1, data = Bids)
+#'
+#' ## Compute its AIC and BIC
 #' AIC(fit)
 #' @export
 AIC.glm_CMP <- function(object, ..., k = 2) {
@@ -78,11 +81,14 @@ print.glm_CMP <- function (x, digits = max(3L, getOption("digits") - 3L), ...) {
 #' @return A vector with the prediction means.
 #'
 #' @examples
+#' ## Fit a model
 #' Bids$size.sq <- Bids$size ^ 2
 #' fit <- glm.CMP(formula.mu = numbids ~ leglrest + rearest + finrest +
 #'                whtknght + bidprem + insthold + size + size.sq + regulatn,
 #'                formula.nu = numbids ~ 1, data = Bids)
-#' predict(fit)
+#'
+#' ## As the newdata parameter is not used the fitted values are obtained
+#' predict(fit, type = "response")
 #' @export
 predict.glm_CMP <- function(object, newdata = NULL,
                            type = c("link", "response"), ...) {
@@ -129,15 +135,22 @@ coef.glm_CMP <- function(object, ...) {
 #' object.
 #'
 #' @param object a fitted object of class inheriting from \code{"glm_CMP"}.
-#' @inheritParams stats::confint
+#' @param parm a specification of which parameters are to be given confidence
+#'   intervals, either a vector of numbers or a vector of names. If missing, all
+#'   parameters are considered.
+#' @param level the confidence level required.
+#' @param ... additional argument(s) for methods.
 #' @return A matrix (or vector) with columns giving lower and upper confidence
 #'   limits for each \code{beta} parameter. These will be labelled as
 #'   (1-level)/2 and 1 - (1-level)/2 in (by default 2.5\% and 97.5\%).
 #' @examples
+#' ## Estimate the model
 #' Bids$size.sq <- Bids$size ^ 2
 #' fit <- glm.CMP(formula.mu = numbids ~ leglrest + rearest + finrest +
 #'                whtknght + bidprem + insthold + size + size.sq + regulatn,
 #'                formula.nu = numbids ~ 1, data = Bids)
+#'
+#' ## Compute confidence intervals
 #' confint(fit)
 #' @export
 confint.glm_CMP <- function (object, parm, level = 0.95, ...)